Tidy up readers and writers a little bit
