+++ pkg/storage/memory/media_bucket.go
@@ -1,0 +1,5 @@
+package memory
+
+type mediaBucket struct {
+	containers []*mediaContainer
+}
